+++ R/ds.R
@@ -20,7 +20,7 @@
 #'  \item{DSSTDY}{Study Day of Start of Disposition Event}
 #'  \item{DSSCAT}{Subcategory for Disposition Event}
 #' }
-#' @seealso \code{\link{ds}} \code{\link{mh}}
+#' @seealso \code{\link{ae}} \code{\link{ds}} \code{\link{dv}} \code{\link{ho}} \code{\link{ie}} \code{\link{mh}} \code{\link{suppho}} \code{\link{ts}}
 #' @keywords datasets ds
 #' @name ds
 #' @examples