+++ R/dv.R
@@ -1,0 +1,21 @@
+#' @title dv
+#'
+#' @description dv created from scratch
+#'
+#' @format A data frame with 75 rows and 8 variables:
+#' \describe{
+#'  \item{STUDYID}{Study Identifier}
+#'  \item{USUBJID}{Unique Subject Identifier}
+#'  \item{DOMAIN}{Domain Abbreviation}
+#'  \item{DVSEQ}{Sequence Number}
+#'  \item{DVTERM}{Protocol Deviation Term}
+#'  \item{DVDECOD}{Standardized Deviation Term}
+#'  \item{DVCAT}{Protocol Deviation Category}
+#'  \item{DVSTDTC}{Start Date/Time of Protocol Deviation}
+#' }
+#' @seealso \code{\link{ae}} \code{\link{ds}} \code{\link{dv}} \code{\link{ho}} \code{\link{ie}} \code{\link{mh}} \code{\link{suppho}} \code{\link{ts}}
+#' @keywords datasets dv
+#' @name dv
+#' @examples
+#' head(data("dv"))
+"dv"