+++ R/ho.R
@@ -1,0 +1,26 @@
+#' @title ho
+#'
+#' @description ho created from scratch
+#'
+#' @format A data frame with 1191 rows and 13 variables:
+#' \describe{
+#'  \item{STUDYID}{Study Identifier}
+#'  \item{DOMAIN}{Domain Abbreviation}
+#'  \item{USUBJID}{Unique Subject Identifier}
+#'  \item{HOSEQ}{Sequence Number}
+#'  \item{HOTERM}{Healthcare Encounter Term}
+#'  \item{HOSTDTC}{Start Date/Time of Healthcare Encounter}
+#'  \item{HOSTDY}{Study Day of Start of Encounter}
+#'  \item{HOENDTC}{End Date/Time of Healthcare Encounter}
+#'  \item{HOENDY}{Study Day of End of Healthcare Encounter}
+#'  \item{HOENRTPT}{End Relative to Reference Time Point}
+#'  \item{HOENRF}{End Relative to Reference Period}
+#'  \item{HODUR}{Duration of Healthcare Encounter}
+#'  \item{HOOCCUR}{Healthcare Encounter Occurrence}
+#' }
+#' @seealso \code{\link{ae}} \code{\link{ds}} \code{\link{dv}} \code{\link{ho}} \code{\link{ie}} \code{\link{mh}} \code{\link{suppho}} \code{\link{ts}}
+#' @keywords datasets ho
+#' @name ho
+#' @examples
+#' head(data("ho"))
+"ho"